requirements.txt
Tkinter
Pillow
IMDbPY